일정 관리 앱 Develop 과제 제작 "Lv4 진행중 HttpSession session"(필수기준)

+++ src/main/java/com/example/schedule2app2develop/schedule/threelayer/ScheduleController.java
@@ -1,6 +1,7 @@
 package com.example.schedule2app2develop.schedule.threelayer;
 
 import com.example.schedule2app2develop.schedule.dto.*;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,7 +16,15 @@ public class ScheduleController {
     private final ScheduleService sdService;
 
     @PostMapping("/schedules")
-    public ResponseEntity<SdCreateResponse> create(@RequestBody SdCreateRequest request) {
+    public ResponseEntity<SdCreateResponse> create(
+            @RequestBody SdCreateRequest request,
+            HttpSession session
+    ) {
+        Long loginUserId = (Long) session.getAttribute("LOGIN_USER");
+
+        if (loginUserId == null) {
+            throw new IllegalStateException("로그인이 필요합니다.");
+        }
         return ResponseEntity.status(HttpStatus.CREATED).body(sdService.create(request));
     }
 
@@ -31,12 +40,26 @@ public class ScheduleController {
     }
 
     @PutMapping("/schedules/{scheduleId}")
-    public ResponseEntity<SdUpdateResponse> update(@PathVariable Long scheduleId, @RequestBody SdUpdateRequest request) {
+    public ResponseEntity<SdUpdateResponse> update(
+            @PathVariable Long scheduleId,
+            @RequestBody SdUpdateRequest request,
+            HttpSession session
+    ) {
+        Long loginUserId = (Long) session.getAttribute("LOGIN_USER");
+
+        if (loginUserId == null) {
+            throw new IllegalStateException("로그인이 필요합니다.");
+        }
         return  ResponseEntity.status(HttpStatus.OK).body(sdService.update(scheduleId, request));
     }
 
     @DeleteMapping("/schedules/{scheduleId}")
-    public ResponseEntity<Void> sdDelete(@PathVariable Long scheduleId) {
+    public ResponseEntity<Void> sdDelete(@PathVariable Long scheduleId, HttpSession session) {
+        Long loginUserId = (Long) session.getAttribute("LOGIN_USER");
+
+        if (loginUserId == null) {
+            throw new IllegalStateException("로그인이 필요합니다.");
+        }
         sdService.delete(scheduleId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
